Correcciones tipo de dato Double a Long pruebas CRUD listas

+++ src/main/java/com/backend/tienda/entidades/DetalleVentas.java
@@ -18,7 +18,7 @@ import lombok.NoArgsConstructor;
 
 public class DetalleVentas {
     @Id
-    private Double codigo_detalle_venta;
+    private Long codigo_detalle_venta;
 
     private int cantidad_producto;
 
@@ -30,7 +30,7 @@ public class DetalleVentas {
     @JoinColumn(name = "codigo_venta")
     private Ventas codigo_venta;
 
-    private double valor_total;
-    private double valor_venta;
-    private double valoriva;
+    private Long valor_total;
+    private Long valor_venta;
+    private Long valoriva;
 }
